created common helper file for shared function in resolution detection processes. adjusted calculation of error for resolution detection to be relative to the predicted unit
